LayoutTests/inspector/console/message-stack-trace-expected.txt

CONSOLE MESSAGE: error message
CONSOLE MESSAGE: TypeError: undefined is not an object (evaluating '[].x.x')
CONSOLE MESSAGE: Unhandled Promise Rejection: rejection-1
CONSOLE MESSAGE: Unhandled Promise Rejection: rejection-2
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating '[].x.x')
Test for the Console.stackTrace in different cases.


== Running test suite: Console.StackTrace
-- Running test case: Console.StackTrace.ConsoleTrace
CALL STACK:
0: [F] (anonymous function)
1: [F] triggerConsoleTrace
ASYNC CALL STACK:
0: --- setTimeout ---
1: [P] Global Code

-- Running test case: Console.StackTrace.ConsoleError
CALL STACK:
0: [F] (anonymous function)
1: [F] triggerConsoleError
ASYNC CALL STACK:
0: --- setTimeout ---
1: [P] Global Code

-- Running test case: Console.StackTrace.UncaughtException
Uncaught exception in test page: TypeError: undefined is not an object (evaluating '[].x.x') [message-stack-trace.html:24]
CALL STACK:
0: [F] (anonymous function)
1: [F] triggerUncaughtException
ASYNC CALL STACK:
0: --- setTimeout ---
1: [P] Global Code

-- Running test case: Console.StackTrace.UnhandledPromiseRejection.PromiseReject
CALL STACK:
0: [N] reject
1: [F] triggerUnhandledRejectionPromiseReject
ASYNC CALL STACK:
0: --- setTimeout ---
1: [P] Global Code

-- Running test case: Console.StackTrace.UnhandledPromiseRejection.ExplicitReject
CALL STACK:
0: [F] (anonymous function)
1: [N] Promise
2: [F] triggerUnhandledRejectionExplicit
ASYNC CALL STACK:
0: --- setTimeout ---
1: [P] Global Code

-- Running test case: Console.StackTrace.UnhandledPromiseRejection.ImplicitReject
CALL STACK:
0: [F] (anonymous function)
ASYNC CALL STACK:
0: --- enqueueJob ---
1: [N] then
2: [F] triggerUnhandledRejectionImplicitWithException
